(wip) refactor: split manifest files
