+++ proofs/src/program/manifest.rs
@@ -33,7 +33,6 @@ use ff::Field;
 use num_bigint::BigInt;
 use serde::{Deserialize, Serialize};
 use serde_json::{json, Value};
-use tiny_keccak::{Hasher, Keccak};
 use tls_client2::CipherSuiteKey;
 use tracing::debug;
 use web_proof_circuits_witness_generator::{
@@ -50,7 +49,8 @@ use crate::{
   circuits::MAX_STACK_HEIGHT,
   program::{
     data::{CircuitData, FoldInput},
-    http::{ManifestRequest, ManifestResponse},
+    manifest,
+    plain_manifest::Manifest,
     F, G1,
   },
   ProofError,
@@ -64,75 +64,6 @@ pub const PUBLIC_IO_VARS: usize = 11;
 pub(crate) const MAX_HTTP_HEADERS: usize = 25;
 /// HTTP/1.1
 pub const HTTP_1_1: &str = "HTTP/1.1";
-
-/// Manifest containing [`ManifestRequest`] and [`ManifestResponse`]
-#[derive(Debug, Clone, Serialize, Deserialize, PartialEq, From)]
-// #[serde(rename_all = "camelCase")]
-pub struct Manifest {
-  // /// Manifest version
-  // pub manifest_version: String,
-  // /// ID of the manifest
-  // pub id:               String,
-  // /// Title of the manifest
-  // pub title:            String,
-  // /// Description of the manifest
-  // pub description:      String,
-  /// HTTP request lock items
-  pub request:  ManifestRequest,
-  /// HTTP response lock items
-  pub response: ManifestResponse,
-}
-
-impl Manifest {
-  /// Validates `Manifest` request and response fields. They are validated against valid statuses,
-  /// http methods, and template variables.
-  pub fn validate(&self) -> Result<(), ProofError> {
-    // TODO: Validate manifest version, id, title, description, prepareUrl
-    self.request.validate()?;
-    self.response.validate()?;
-
-    // TODO: Do we want to run this here instead of in Request::validate in order to cross-examine
-    //  Request and Response template vars?
-    // TODO: Commented out because it breaks when existing fixtures, i.e. tokens encoded in the
-    //  body of client.tee_tcp_local.json
-    // self.request.validate_vars()?;
-    Ok(())
-  }
-
-  /// Compute a `Keccak256` hash of the serialized Manifest
-  pub fn to_keccak_digest(&self) -> Result<[u8; 32], ProofError> {
-    let as_bytes: Vec<u8> = self.try_into()?;
-    let mut hasher = Keccak::v256();
-    let mut output = [0u8; 32];
-
-    hasher.update(&as_bytes);
-    hasher.finalize(&mut output);
-
-    Ok(output)
-  }
-}
-
-impl TryFrom<&[u8]> for Manifest {
-  type Error = serde_json::Error;
-
-  fn try_from(bytes: &[u8]) -> Result<Self, Self::Error> { serde_json::from_slice(bytes) }
-}
-
-impl TryFrom<&Manifest> for Vec<u8> {
-  type Error = serde_json::Error;
-
-  fn try_from(manifest: &Manifest) -> Result<Self, Self::Error> { serde_json::to_vec(manifest) }
-}
-
-impl TryFrom<Manifest> for Vec<u8> {
-  type Error = serde_json::Error;
-
-  fn try_from(manifest: Manifest) -> Result<Self, Self::Error> { serde_json::to_vec(&manifest) }
-}
-
-impl From<OrigoManifest> for Manifest {
-  fn from(value: OrigoManifest) -> Self { value.0 }
-}
 
 /// A type of `Manifest` with special methods for Origo mode.
 #[derive(Debug, Clone, Serialize, Deserialize, From)]
@@ -151,6 +82,252 @@ impl TryFrom<&OrigoManifest> for Vec<u8> {
 
   fn try_from(manifest: &OrigoManifest) -> Result<Self, Self::Error> {
     serde_json::to_vec(&manifest.0)
+  }
+}
+
+impl OrigoManifest {
+  /// Initial inputs
+  pub fn initial_inputs<const MAX_STACK_HEIGHT: usize, const CIRCUIT_SIZE: usize>(
+    &self,
+    // TODO (Sambhav): can remove copying
+    request_ciphertext: &[Vec<u8>],
+    response_ciphertext: &[Vec<u8>],
+  ) -> Result<InitialNIVCInputs, ProofError> {
+    let ciphertext_digest =
+      manifest::compute_ciphertext_digest::<CIRCUIT_SIZE>(request_ciphertext, response_ciphertext);
+
+    // TODO: This assumes the start line format here as well.
+    // digest the start line for request/response using the ciphertext_digest as a random input
+    let request_start_line_bytes =
+      format!("{} {} {}", &self.0.request.method, &self.0.request.url, &self.0.request.version);
+    let request_start_line_digest =
+      polynomial_digest(request_start_line_bytes.as_bytes(), ciphertext_digest, 0);
+    debug!(
+      "WITNESS (request): start_line_digest={:?}, hex={:?}",
+      request_start_line_digest,
+      hex::encode(request_start_line_digest.to_bytes())
+    );
+
+    let response_start_line_bytes = format!(
+      "{} {} {}",
+      &self.0.response.version, &self.0.response.status, &self.0.response.message
+    );
+    let response_start_line_digest =
+      polynomial_digest(response_start_line_bytes.as_bytes(), ciphertext_digest, 0);
+
+    // Digest all the headers
+    let request_header_bytes = headers_to_bytes(&self.0.request.headers);
+    let request_headers_digest =
+      request_header_bytes.map(|bytes| polynomial_digest(&bytes, ciphertext_digest, 0));
+    // debug!(
+    //   "WITNESS (request): headers_digest={:?}, hex={:?}",
+    //   request_headers_digest,
+    //   request_headers_digest.clone().into_iter().map(|f| hex::encode(f.to_bytes()))
+    // );
+
+    let response_header_bytes = headers_to_bytes(&self.0.response.headers);
+    let response_headers_digest =
+      response_header_bytes.map(|bytes| polynomial_digest(&bytes, ciphertext_digest, 0));
+    // debug!(
+    //   "WITNESS (response): headers_digest={:?}, hex={:?}",
+    //   response_headers_digest,
+    //   response_headers_digest.clone().into_iter().map(|f| hex::encode(f.to_bytes()))
+    // );
+
+    // Digest the JSON sequence
+    let raw_response_json_machine =
+      RawJsonMachine::<MAX_STACK_HEIGHT>::from_chosen_sequence_and_input(
+        ciphertext_digest,
+        &self.0.response.body.json_path,
+      )?;
+    let json_sequence_digest_hash =
+      poseidon::<1>(&[raw_response_json_machine.compress_tree_hash()]);
+
+    // push request/response start line digest and headers digest into one all_digest vec
+    let mut all_digest = vec![request_start_line_digest, response_start_line_digest];
+    all_digest.extend(request_headers_digest);
+    all_digest.extend(response_headers_digest);
+
+    let header_verification_lock = all_digest.iter().map(|d| poseidon::<1>(&[*d])).sum::<F<G1>>();
+    let num_matches = 1 + self.0.request.headers.len() + 1 + self.0.response.headers.len();
+
+    let initial_http_machine_digest =
+      polynomial_digest(&[1, 0, 0, 0, 0, 0, 0, 1], ciphertext_digest, 0);
+
+    Ok(InitialNIVCInputs {
+      ciphertext_digest,
+      initial_nivc_input: [
+        ciphertext_digest,
+        F::<G1>::ONE,
+        F::<G1>::ONE,
+        initial_http_machine_digest,
+        header_verification_lock,
+        F::<G1>::from(num_matches as u64),
+        F::<G1>::ZERO,
+        F::<G1>::ONE,
+        F::<G1>::ZERO,
+        json_sequence_digest_hash,
+        F::<G1>::ZERO,
+      ],
+      headers_digest: all_digest,
+    })
+  }
+
+  /// Builds inputs
+  pub fn build_inputs<const CIRCUIT_SIZE: usize>(
+    &self,
+    request_inputs: &EncryptionInput,
+    response_inputs: &EncryptionInput,
+  ) -> Result<NivcCircuitInputs, ProofError> {
+    assert_eq!(request_inputs.plaintext.len(), request_inputs.ciphertext.len());
+    assert_eq!(response_inputs.plaintext.len(), response_inputs.ciphertext.len());
+
+    let mut private_inputs = vec![];
+    // TODO: Fold inputs are actually not modified, ever - We return en empty HashMap
+    let fold_inputs: HashMap<String, FoldInput> = HashMap::new();
+
+    let InitialNIVCInputs { ciphertext_digest, initial_nivc_input, headers_digest } =
+      self.initial_inputs::<MAX_STACK_HEIGHT, CIRCUIT_SIZE>(
+        &request_inputs.ciphertext,
+        &response_inputs.ciphertext,
+      )?;
+
+    let _ = manifest::build_plaintext_authentication_circuit_inputs::<CIRCUIT_SIZE>(
+      request_inputs,
+      ciphertext_digest,
+      &mut private_inputs,
+    )?;
+    // debug!("private_inputs: {:?}", private_inputs.len());
+
+    let (_, request_body) = manifest::build_http_verification_circuit_inputs::<CIRCUIT_SIZE>(
+      &request_inputs.plaintext,
+      ciphertext_digest,
+      &headers_digest,
+      &mut private_inputs,
+    )?;
+    // debug!("private_inputs: {:?}", private_inputs.len());
+
+    let _ = manifest::build_json_extraction_circuit_inputs::<CIRCUIT_SIZE>(
+      &request_body,
+      ciphertext_digest,
+      (Some(&[]), Some(&self.0.response.body.json_path)), /* WARN: sending response keys for
+                                                           * request */
+      &mut private_inputs,
+    )?;
+    // debug!("private_inputs: {:?}", private_inputs.len());
+
+    let _ = manifest::build_plaintext_authentication_circuit_inputs::<CIRCUIT_SIZE>(
+      response_inputs,
+      ciphertext_digest,
+      &mut private_inputs,
+    )?;
+
+    // debug!("private_inputs: {:?}", private_inputs.len());
+    let (_, response_body) = manifest::build_http_verification_circuit_inputs::<CIRCUIT_SIZE>(
+      &response_inputs.plaintext,
+      ciphertext_digest,
+      &headers_digest,
+      &mut private_inputs,
+    )?;
+
+    let _ = manifest::build_json_extraction_circuit_inputs::<CIRCUIT_SIZE>(
+      &response_body,
+      ciphertext_digest,
+      (None, Some(&self.0.response.body.json_path)),
+      &mut private_inputs,
+    )?;
+
+    Ok(NivcCircuitInputs { private_inputs, fold_inputs, initial_nivc_input })
+  }
+
+  /// Builds ROM for [`Manifest`] request and response.
+  pub fn build_rom<const CIRCUIT_SIZE: usize>(
+    &self,
+    request_inputs: &EncryptionInput,
+    response_inputs: &EncryptionInput,
+  ) -> NIVCRom {
+    let plaintext_authentication_label = String::from("PLAINTEXT_AUTHENTICATION");
+    let http_verification_label = String::from("HTTP_VERIFICATION");
+    let json_extraction_label = String::from("JSON_EXTRACTION");
+
+    let mut rom = vec![];
+    // ------------------- Request -------------------
+
+    let combined_request_plaintext_length: usize =
+      request_inputs.plaintext.iter().map(|x| x.len()).sum();
+
+    // plaintext_authentication_label is duplicated `response_packets` times
+    let mut rom_data = HashMap::new();
+    let mut plaintext_circuit_counter = 0;
+    for c in request_inputs.ciphertext.iter() {
+      let circuit_count = (c.len() as f64 / CIRCUIT_SIZE as f64).ceil() as usize;
+      for _ in 0..circuit_count {
+        let plaintext_circuit =
+          format!("{}_{}", plaintext_authentication_label, plaintext_circuit_counter);
+        rom_data.insert(plaintext_circuit.clone(), CircuitData { opcode: 0 });
+        rom.push(plaintext_circuit);
+        plaintext_circuit_counter += 1;
+      }
+    }
+    // calculate number of circuits required for request, i.e. ceil(length/CIRCUIT_SIZE)
+    let request_circuit_count =
+      (combined_request_plaintext_length as f64 / CIRCUIT_SIZE as f64).ceil() as usize;
+    (0..request_circuit_count).for_each(|i| {
+      let http_circuit = format!("{}_{}", http_verification_label, i);
+      rom_data.insert(http_circuit.clone(), CircuitData { opcode: 1 });
+      rom.push(http_circuit);
+    });
+
+    let combined_request = request_inputs.plaintext.iter().flatten().cloned().collect::<Vec<u8>>();
+    let request_body = compute_http_witness(&combined_request, HttpMaskType::Body);
+    let request_json_circuit_count =
+      (request_body.len() as f64 / CIRCUIT_SIZE as f64).ceil() as usize;
+
+    (0..request_json_circuit_count).for_each(|i| {
+      let json_circuit = format!("{}_{}", json_extraction_label, i);
+      rom_data.insert(json_circuit.clone(), CircuitData { opcode: 2 });
+      rom.push(json_circuit);
+    });
+
+    // ------------------- Response -------------------
+
+    let combined_response_plaintext_length: usize =
+      response_inputs.plaintext.iter().map(|x| x.len()).sum();
+
+    // plaintext_authentication_label is duplicated `response_packets` times
+    for c in response_inputs.ciphertext.iter() {
+      let circuit_count = (c.len() as f64 / CIRCUIT_SIZE as f64).ceil() as usize;
+      for _ in 0..circuit_count {
+        let plaintext_circuit =
+          format!("{}_{}", plaintext_authentication_label, plaintext_circuit_counter);
+        rom_data.insert(plaintext_circuit.clone(), CircuitData { opcode: 0 });
+        rom.push(plaintext_circuit);
+        plaintext_circuit_counter += 1;
+      }
+    }
+
+    // calculate number of circuits required for response, i.e. ceil(length/CIRCUIT_SIZE)
+    let response_circuit_count =
+      (combined_response_plaintext_length as f64 / CIRCUIT_SIZE as f64).ceil() as usize;
+    (0..response_circuit_count).for_each(|i| {
+      let http_circuit = format!("{}_{}", http_verification_label, i + request_circuit_count);
+      rom_data.insert(http_circuit.clone(), CircuitData { opcode: 1 });
+      rom.push(http_circuit);
+    });
+
+    let combined_response =
+      response_inputs.plaintext.iter().flatten().cloned().collect::<Vec<u8>>();
+    let response_body = compute_http_witness(&combined_response, HttpMaskType::Body);
+    let response_json_circuit_count =
+      (response_body.len() as f64 / CIRCUIT_SIZE as f64).ceil() as usize;
+
+    (0..response_json_circuit_count).for_each(|i| {
+      let json_circuit = format!("{}_{}", json_extraction_label, i + request_json_circuit_count);
+      rom_data.insert(json_circuit.clone(), CircuitData { opcode: 2 });
+      rom.push(json_circuit);
+    });
+
+    NIVCRom { circuit_data: rom_data, rom }
   }
 }
 
@@ -508,261 +685,12 @@ pub fn compute_ciphertext_digest<const CIRCUIT_SIZE: usize>(
   ciphertext_digest
 }
 
-impl OrigoManifest {
-  /// Initial inputs
-  pub fn initial_inputs<const MAX_STACK_HEIGHT: usize, const CIRCUIT_SIZE: usize>(
-    &self,
-    // TODO (Sambhav): can remove copying
-    request_ciphertext: &[Vec<u8>],
-    response_ciphertext: &[Vec<u8>],
-  ) -> Result<InitialNIVCInputs, ProofError> {
-    let ciphertext_digest =
-      compute_ciphertext_digest::<CIRCUIT_SIZE>(request_ciphertext, response_ciphertext);
-
-    // TODO: This assumes the start line format here as well.
-    // digest the start line for request/response using the ciphertext_digest as a random input
-    let request_start_line_bytes =
-      format!("{} {} {}", &self.0.request.method, &self.0.request.url, &self.0.request.version);
-    let request_start_line_digest =
-      polynomial_digest(request_start_line_bytes.as_bytes(), ciphertext_digest, 0);
-    debug!(
-      "WITNESS (request): start_line_digest={:?}, hex={:?}",
-      request_start_line_digest,
-      hex::encode(request_start_line_digest.to_bytes())
-    );
-
-    let response_start_line_bytes = format!(
-      "{} {} {}",
-      &self.0.response.version, &self.0.response.status, &self.0.response.message
-    );
-    let response_start_line_digest =
-      polynomial_digest(response_start_line_bytes.as_bytes(), ciphertext_digest, 0);
-
-    // Digest all the headers
-    let request_header_bytes = headers_to_bytes(&self.0.request.headers);
-    let request_headers_digest =
-      request_header_bytes.map(|bytes| polynomial_digest(&bytes, ciphertext_digest, 0));
-    // debug!(
-    //   "WITNESS (request): headers_digest={:?}, hex={:?}",
-    //   request_headers_digest,
-    //   request_headers_digest.clone().into_iter().map(|f| hex::encode(f.to_bytes()))
-    // );
-
-    let response_header_bytes = headers_to_bytes(&self.0.response.headers);
-    let response_headers_digest =
-      response_header_bytes.map(|bytes| polynomial_digest(&bytes, ciphertext_digest, 0));
-    // debug!(
-    //   "WITNESS (response): headers_digest={:?}, hex={:?}",
-    //   response_headers_digest,
-    //   response_headers_digest.clone().into_iter().map(|f| hex::encode(f.to_bytes()))
-    // );
-
-    // Digest the JSON sequence
-    let raw_response_json_machine =
-      RawJsonMachine::<MAX_STACK_HEIGHT>::from_chosen_sequence_and_input(
-        ciphertext_digest,
-        &self.0.response.body.json_path,
-      )?;
-    let json_sequence_digest_hash =
-      poseidon::<1>(&[raw_response_json_machine.compress_tree_hash()]);
-
-    // push request/response start line digest and headers digest into one all_digest vec
-    let mut all_digest = vec![request_start_line_digest, response_start_line_digest];
-    all_digest.extend(request_headers_digest);
-    all_digest.extend(response_headers_digest);
-
-    let header_verification_lock = all_digest.iter().map(|d| poseidon::<1>(&[*d])).sum::<F<G1>>();
-    let num_matches = 1 + self.0.request.headers.len() + 1 + self.0.response.headers.len();
-
-    let initial_http_machine_digest =
-      polynomial_digest(&[1, 0, 0, 0, 0, 0, 0, 1], ciphertext_digest, 0);
-
-    Ok(InitialNIVCInputs {
-      ciphertext_digest,
-      initial_nivc_input: [
-        ciphertext_digest,
-        F::<G1>::ONE,
-        F::<G1>::ONE,
-        initial_http_machine_digest,
-        header_verification_lock,
-        F::<G1>::from(num_matches as u64),
-        F::<G1>::ZERO,
-        F::<G1>::ONE,
-        F::<G1>::ZERO,
-        json_sequence_digest_hash,
-        F::<G1>::ZERO,
-      ],
-      headers_digest: all_digest,
-    })
-  }
-
-  /// Builds inputs
-  pub fn build_inputs<const CIRCUIT_SIZE: usize>(
-    &self,
-    request_inputs: &EncryptionInput,
-    response_inputs: &EncryptionInput,
-  ) -> Result<NivcCircuitInputs, ProofError> {
-    assert_eq!(request_inputs.plaintext.len(), request_inputs.ciphertext.len());
-    assert_eq!(response_inputs.plaintext.len(), response_inputs.ciphertext.len());
-
-    let mut private_inputs = vec![];
-    // TODO: Fold inputs are actually not modified, ever - We return en empty HashMap
-    let fold_inputs: HashMap<String, FoldInput> = HashMap::new();
-
-    let InitialNIVCInputs { ciphertext_digest, initial_nivc_input, headers_digest } =
-      self.initial_inputs::<MAX_STACK_HEIGHT, CIRCUIT_SIZE>(
-        &request_inputs.ciphertext,
-        &response_inputs.ciphertext,
-      )?;
-
-    let _ = build_plaintext_authentication_circuit_inputs::<CIRCUIT_SIZE>(
-      request_inputs,
-      ciphertext_digest,
-      &mut private_inputs,
-    )?;
-    // debug!("private_inputs: {:?}", private_inputs.len());
-
-    let (_, request_body) = build_http_verification_circuit_inputs::<CIRCUIT_SIZE>(
-      &request_inputs.plaintext,
-      ciphertext_digest,
-      &headers_digest,
-      &mut private_inputs,
-    )?;
-    // debug!("private_inputs: {:?}", private_inputs.len());
-
-    let _ = build_json_extraction_circuit_inputs::<CIRCUIT_SIZE>(
-      &request_body,
-      ciphertext_digest,
-      (Some(&[]), Some(&self.0.response.body.json_path)), // WARN: sending response keys for request
-      &mut private_inputs,
-    )?;
-    // debug!("private_inputs: {:?}", private_inputs.len());
-
-    let _ = build_plaintext_authentication_circuit_inputs::<CIRCUIT_SIZE>(
-      response_inputs,
-      ciphertext_digest,
-      &mut private_inputs,
-    )?;
-
-    // debug!("private_inputs: {:?}", private_inputs.len());
-    let (_, response_body) = build_http_verification_circuit_inputs::<CIRCUIT_SIZE>(
-      &response_inputs.plaintext,
-      ciphertext_digest,
-      &headers_digest,
-      &mut private_inputs,
-    )?;
-
-    let _ = build_json_extraction_circuit_inputs::<CIRCUIT_SIZE>(
-      &response_body,
-      ciphertext_digest,
-      (None, Some(&self.0.response.body.json_path)),
-      &mut private_inputs,
-    )?;
-
-    Ok(NivcCircuitInputs { private_inputs, fold_inputs, initial_nivc_input })
-  }
-
-  /// Builds ROM for [`Manifest`] request and response.
-  pub fn build_rom<const CIRCUIT_SIZE: usize>(
-    &self,
-    request_inputs: &EncryptionInput,
-    response_inputs: &EncryptionInput,
-  ) -> NIVCRom {
-    let plaintext_authentication_label = String::from("PLAINTEXT_AUTHENTICATION");
-    let http_verification_label = String::from("HTTP_VERIFICATION");
-    let json_extraction_label = String::from("JSON_EXTRACTION");
-
-    let mut rom = vec![];
-    // ------------------- Request -------------------
-
-    let combined_request_plaintext_length: usize =
-      request_inputs.plaintext.iter().map(|x| x.len()).sum();
-
-    // plaintext_authentication_label is duplicated `response_packets` times
-    let mut rom_data = HashMap::new();
-    let mut plaintext_circuit_counter = 0;
-    for c in request_inputs.ciphertext.iter() {
-      let circuit_count = (c.len() as f64 / CIRCUIT_SIZE as f64).ceil() as usize;
-      for _ in 0..circuit_count {
-        let plaintext_circuit =
-          format!("{}_{}", plaintext_authentication_label, plaintext_circuit_counter);
-        rom_data.insert(plaintext_circuit.clone(), CircuitData { opcode: 0 });
-        rom.push(plaintext_circuit);
-        plaintext_circuit_counter += 1;
-      }
-    }
-    // calculate number of circuits required for request, i.e. ceil(length/CIRCUIT_SIZE)
-    let request_circuit_count =
-      (combined_request_plaintext_length as f64 / CIRCUIT_SIZE as f64).ceil() as usize;
-    (0..request_circuit_count).for_each(|i| {
-      let http_circuit = format!("{}_{}", http_verification_label, i);
-      rom_data.insert(http_circuit.clone(), CircuitData { opcode: 1 });
-      rom.push(http_circuit);
-    });
-
-    let combined_request = request_inputs.plaintext.iter().flatten().cloned().collect::<Vec<u8>>();
-    let request_body = compute_http_witness(&combined_request, HttpMaskType::Body);
-    let request_json_circuit_count =
-      (request_body.len() as f64 / CIRCUIT_SIZE as f64).ceil() as usize;
-
-    (0..request_json_circuit_count).for_each(|i| {
-      let json_circuit = format!("{}_{}", json_extraction_label, i);
-      rom_data.insert(json_circuit.clone(), CircuitData { opcode: 2 });
-      rom.push(json_circuit);
-    });
-
-    // ------------------- Response -------------------
-
-    let combined_response_plaintext_length: usize =
-      response_inputs.plaintext.iter().map(|x| x.len()).sum();
-
-    // plaintext_authentication_label is duplicated `response_packets` times
-    for c in response_inputs.ciphertext.iter() {
-      let circuit_count = (c.len() as f64 / CIRCUIT_SIZE as f64).ceil() as usize;
-      for _ in 0..circuit_count {
-        let plaintext_circuit =
-          format!("{}_{}", plaintext_authentication_label, plaintext_circuit_counter);
-        rom_data.insert(plaintext_circuit.clone(), CircuitData { opcode: 0 });
-        rom.push(plaintext_circuit);
-        plaintext_circuit_counter += 1;
-      }
-    }
-
-    // calculate number of circuits required for response, i.e. ceil(length/CIRCUIT_SIZE)
-    let response_circuit_count =
-      (combined_response_plaintext_length as f64 / CIRCUIT_SIZE as f64).ceil() as usize;
-    (0..response_circuit_count).for_each(|i| {
-      let http_circuit = format!("{}_{}", http_verification_label, i + request_circuit_count);
-      rom_data.insert(http_circuit.clone(), CircuitData { opcode: 1 });
-      rom.push(http_circuit);
-    });
-
-    let combined_response =
-      response_inputs.plaintext.iter().flatten().cloned().collect::<Vec<u8>>();
-    let response_body = compute_http_witness(&combined_response, HttpMaskType::Body);
-    let response_json_circuit_count =
-      (response_body.len() as f64 / CIRCUIT_SIZE as f64).ceil() as usize;
-
-    (0..response_json_circuit_count).for_each(|i| {
-      let json_circuit = format!("{}_{}", json_extraction_label, i + request_json_circuit_count);
-      rom_data.insert(json_circuit.clone(), CircuitData { opcode: 2 });
-      rom.push(json_circuit);
-    });
-
-    NIVCRom { circuit_data: rom_data, rom }
-  }
-}
-
 #[cfg(test)]
 mod tests {
   use super::*;
-  use crate::{
-    program::http::{ManifestResponseBody, TemplateVar},
-    request, response,
-    tests::inputs::{
-      complex_manifest, complex_request_inputs, complex_response_inputs, simple_request_inputs,
-      simple_response_inputs, TEST_MANIFEST,
-    },
+  use crate::tests::inputs::{
+    complex_manifest, complex_request_inputs, complex_response_inputs, simple_request_inputs,
+    simple_response_inputs, TEST_MANIFEST,
   };
 
   #[test]
@@ -892,9 +820,9 @@ mod tests {
 
   #[test]
   fn test_rom_from_inputs() {
-    let manifest: OrigoManifest = serde_json::from_str(TEST_MANIFEST).unwrap();
+    let manifest: Manifest = serde_json::from_str(TEST_MANIFEST).unwrap();
     let (request_inputs, response_inputs) = simple_inputs();
-    assert_rom_from_inputs::<512>(manifest, request_inputs, response_inputs);
+    assert_rom_from_inputs::<512>(manifest.into(), request_inputs, response_inputs);
 
     let (request_inputs, response_inputs) = complex_inputs();
     assert_rom_from_inputs::<512>(complex_manifest().into(), request_inputs, response_inputs);
@@ -1009,159 +937,5 @@ mod tests {
       0, 1, 0, 1, 0, 0, 0, 0, 0, 1, 0, 1, 0, 1, 0, 0, 0, 1, 0, 1, 0, 1, 0, 0, 0, 1, 0, 0, 1, 0, 0,
       0
     ]]);
-  }
-
-  #[test]
-  fn test_manifest_serialization() {
-    let manifest: Manifest = serde_json::from_str(TEST_MANIFEST).unwrap();
-    let serialized: Vec<u8> = manifest.clone().try_into().unwrap();
-    let deserialized = Manifest::try_from(serialized.as_slice()).unwrap();
-    assert_eq!(manifest, deserialized);
-  }
-
-  macro_rules! create_manifest {
-    (
-        $request:expr,
-        $response:expr
-        $(, $field:ident = $value:expr)* $(,)?
-    ) => {{
-        Manifest {
-            // manifest_version: "1".to_string(),
-            // id: "Default Manifest ID".to_string(),
-            // title: "Default Manifest Title".to_string(),
-            // description: "Default description.".to_string(),
-            request: $request,
-            response: $response,
-            $(
-                $field: $value,
-            )*
-        }
-    }};
-  }
-
-  // TODO: Using this TEST_MANIFEST duplicate here because modifying the original one breaks
-  //  `test_end_to_end_proofs` test with error: thread 'tests::test_end_to_end_proofs_get' panicked
-  // at proofs/src/tests/mod.rs:230:84:  called `Result::unwrap()` on an `Err` value:
-  // NovaError(InvalidSumcheckProof)
-  const TEST_MANIFEST_WITHOUT_VARS: &str = r#"
-{
-    "manifestVersion": "1",
-    "id": "reddit-user-karma",
-    "title": "Total Reddit Karma",
-    "description": "Generate a proof that you have a certain amount of karma",
-    "prepareUrl": "https://www.reddit.com/login/",
-    "request": {
-        "method": "GET",
-        "url": "https://gist.githubusercontent.com/mattes/23e64faadb5fd4b5112f379903d2572e/raw/74e517a60c21a5c11d94fec8b572f68addfade39/example.json",
-        "headers": {
-            "host": "gist.githubusercontent.com",
-            "connection": "close",
-            "Authorization": "Bearer <% token %>",
-            "User-Agent": "test-agent"
-        }
-    },
-    "response": {
-        "status": "200",
-        "headers": {
-            "Content-Type": "text/plain; charset=utf-8",
-            "Content-Length": "22"
-        },
-        "body": {
-            "json": [
-                "hello"
-            ]
-        }
-    }
-}
-"#;
-
-  #[test]
-  fn test_green_path_manifest_validation() {
-    let manifest: Manifest = serde_json::from_str(TEST_MANIFEST).unwrap();
-    let result = manifest.validate();
-    assert!(result.is_ok());
-  }
-
-  #[test]
-  fn test_parse_manifest_without_vars() {
-    let manifest: Manifest = serde_json::from_str(TEST_MANIFEST_WITHOUT_VARS).unwrap();
-    let result = manifest.validate();
-    assert!(result.is_ok());
-
-    assert!(manifest.request.body.is_none()); // Optional field we omitted
-    assert_eq!(manifest.request.vars, HashMap::new()); // Optional field we provide default for
-  }
-
-  #[test]
-  fn test_manifest_validation_invalid_method() {
-    let manifest = create_manifest!(request!(method: "INVALID".to_string()), response!());
-    let result = manifest.validate();
-    assert!(result.is_err());
-    if let Err(ProofError::InvalidManifest(message)) = result {
-      assert_eq!(message, "Invalid HTTP method");
-    } else {
-      panic!("Expected ProofError::InvalidManifest");
-    }
-  }
-
-  #[test]
-  fn test_manifest_validation_invalid_url() {
-    let manifest = create_manifest!(request!(url: "ftp://example.com".to_string()), response!(),);
-    let result = manifest.validate();
-    assert!(result.is_err());
-    if let Err(ProofError::InvalidManifest(message)) = result {
-      assert_eq!(message, "Only HTTPS URLs are allowed");
-    } else {
-      panic!("Expected ProofError::InvalidManifest");
-    }
-  }
-
-  #[test]
-  fn test_manifest_validation_invalid_response_status() {
-    let manifest = create_manifest!(request!(), response!(status: "500".to_string()),);
-    let result = manifest.validate();
-    assert!(result.is_err());
-    if let Err(ProofError::InvalidManifest(message)) = result {
-      assert_eq!(message, "Unsupported HTTP status");
-    } else {
-      panic!("Expected ProofError::InvalidManifest");
-    }
-  }
-
-  #[ignore] // TODO: Unignore after fixtures are fixed and token validation is re-enabled
-  #[test]
-  fn test_manifest_validation_missing_vars() {
-    let mut vars = HashMap::new();
-    vars.insert("TOKEN".to_string(), TemplateVar {
-      regex:  Some("^[A-Za-z0-9]+$".to_string()),
-      length: Some(20),
-      r#type: Some("base64".to_string()),
-    });
-    let manifest = create_manifest!(
-      request!(
-          headers: HashMap::new(), // Invalid because "Authorization" makes use of `<TOKEN>`
-          vars: vars
-      ),
-      response!(),
-    );
-    let result = manifest.validate();
-    assert!(result.is_err());
-  }
-
-  #[test]
-  fn test_manifest_validation_invalid_content_type() {
-    let manifest = create_manifest!(
-      request!(),
-      response!(headers: HashMap::from([
-          ("Content-Type".to_string(), "invalid/type".to_string())
-      ])),
-    );
-    let result = manifest.validate();
-    assert!(result.is_err());
-    if let Err(ProofError::InvalidManifest(message)) = result {
-      assert_eq!(message, "Invalid Content-Type header: invalid/type");
-    } else {
-      panic!("Expected ProofError::InvalidManifest");
-    }
   }
 }
